Alex: Updated hash method
I updated the hash method to be named Hash_djb2 to accredit the writter.  I also modified the method to use uint32_t as the hashing type and not uint64_t as the original method would have a minimum storage of 32 bits due to its usage of unsigned long.

+++ include/reflection/MetaData.h
@@ -29,7 +29,7 @@ public:
 	inline bool Has(HashID id) const;
 
 	// -Types-
-	using Iterator = std::unordered_map<uint64_t, Any>::iterator;
+	using Iterator = std::unordered_map<uint32_t, Any>::iterator;
 
 	// -Iterators-
 
@@ -47,7 +47,7 @@ public:
 
 private:
 
-	std::unordered_map<uint64_t, Any> m_Data;
+	std::unordered_map<uint32_t, Any> m_Data;
 
 };
